refactor: move Colour.hpp to src/Colour/Colour.hpp
